# PERF: Fixed performance regression in Series init