@TomAugspurger 2019-12-30
Also pushed a change to avoid calling `construct_from_string` from `ExtensionDtype.is_dtype` for non-string inputs in the first place.